Split tests for overall dependent management into separate file
