+++ Testing/functional/tests/cypress/support/functions/dependent.js
@@ -1,0 +1,7 @@
+export function getCardSelector(hdid) {
+    return `[data-testid=dependent-card-${hdid}]`;
+}
+
+export function getTabButtonSelector(hdid, tab) {
+    return `#${tab}-tab-button-${hdid}`;
+}
